Split up giant file LfpDisplayCanvas into its constituent classes
